feat(components): add prefix and postfix
